// src/commons/types/generate/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Date: { input: any; output: any };
};

export type IBoardReturn = {
  __typename?: "BoardReturn";
  contents?: Maybe<Scalars["String"]["output"]>;
  createdAt?: Maybe<Scalars["Date"]["output"]>;
  like?: Maybe<Scalars["Int"]["output"]>;
  number?: Maybe<Scalars["Int"]["output"]>;
  title?: Maybe<Scalars["String"]["output"]>;
  writer?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type IProductReturn = {
  __typename?: "ProductReturn";
  _id?: Maybe<Scalars["ID"]["output"]>;
  createdAt?: Maybe<Scalars["Date"]["output"]>;
  detail?: Maybe<Scalars["String"]["output"]>;
  name?: Maybe<Scalars["String"]["output"]>;
  price?: Maybe<Scalars["Int"]["output"]>;
  seller?: Maybe<Scalars["String"]["output"]>;
};

export type IProfileReturn = {
  __typename?: "ProfileReturn";
  age?: Maybe<Scalars["Int"]["output"]>;
  name?: Maybe<Scalars["String"]["output"]>;
  number?: Maybe<Scalars["Int"]["output"]>;
  school?: Maybe<Scalars["String"]["output"]>;
};

export type IQuery = {
  __typename?: "Query";
  fetchBoard?: Maybe<IBoardReturn>;
  fetchBoards?: Maybe<Array<IBoardReturn>>;
  fetchBoardsCount: Scalars["Int"]["output"];
  fetchProduct?: Maybe<IProductReturn>;
  fetchProducts?: Maybe<Array<IProductReturn>>;
  fetchProductsCount: Scalars["Int"]["output"];
  fetchProfile?: Maybe<IProfileReturn>;
  fetchProfiles?: Maybe<Array<IProfileReturn>>;
  fetchProfilesCount: Scalars["Int"]["output"];
};

export type IQueryFetchBoardArgs = {
  number?: InputMaybe<Scalars["Int"]["input"]>;
};

export type IQueryFetchBoardsArgs = {
  page?: InputMaybe<Scalars["Int"]["input"]>;
};